beatbox/src/test/java/com/nanacone/beatbox/SoundViewModelTest.kt
package com.nanacone.beatbox

import org.hamcrest.MatcherAssert
import org.hamcrest.core.Is.`is`
import org.junit.Assert.*

import org.junit.Before
import org.junit.Test
import org.mockito.Mockito.mock
import org.mockito.Mockito.verify

class SoundViewModelTest {

    private lateinit var beatBox: BeatBox
    private lateinit var sound: Sound
    private lateinit var subject: SoundViewModel

    @Before // 각 테스트(@Test)가 진행되기 전에 딱 한번만 실행됨
    fun setUp() {
        beatBox = mock(BeatBox::class.java) // 모의 객체
        sound = Sound("assetPath")
        subject = SoundViewModel(beatBox)
        subject.sound = sound
    }

    @Test
    fun exposesSoundNameAsTitle() {
        MatcherAssert.assertThat(subject.title, `is`(sound.name))
    }

    @Test
    fun callsBeatBoxPlayOnButtonClicked() {
        subject.onButtonClicked()
        verify(beatBox).play(sound) // beatbox 의 play 함수가 호출되었는지 확인
    }
}